// starting-point/libs/workspace-hxp/governance-services-extension/domain-shell/src/lib/search/models/search-options.interface.ts
/*
 * Copyright © 2005 - 2021 Alfresco Software, Ltd. All rights reserved.
 *
 * License rights for this program may be obtained from Alfresco Software, Ltd.
 * pursuant to a written agreement and any use of this program without such an
 * agreement is prohibited.
 */

export interface GovernanceSearchOptions {
    limit?: number;
    exclusiveStartKey?: string;
    sortColumn?: string;
    sortDirection?: 'asc' | 'desc' | '';
}