// starting-point/libs/workspace-hxp/governance-services-extension/domain-shell/src/lib/search/search-results/services/governance-search-result.interface.ts
/*
 * Copyright © 2005 - 2021 Alfresco Software, Ltd. All rights reserved.
 *
 * License rights for this program may be obtained from Alfresco Software, Ltd.
 * pursuant to a written agreement and any use of this program without such an
 * agreement is prohibited.
 */

import { GovernanceRecord } from '../../../mocks/record.type';

export interface GovernanceSearchResult {
    lastEvaluatedKey: string;
    content: GovernanceRecord[];
}